Incognitos/Incognitos CARE (Prototype)/requirements.txt
Flask==1.1.2
gunicorn==20.1.0
google-api-python-client==2.2.0